changed NETWORK_STATISTICS::[BytesInFlight, CongestionWindow] back to uint32_t

+++ src/inc/msquic.h
@@ -643,11 +643,11 @@ typedef struct QUIC_STATISTICS_V2 {
 
 typedef struct NETWORK_STATISTICS
 {
-    uint64_t BytesInFlight;                      // Bytes that were sent on the wire, but not yet acked
+    uint32_t BytesInFlight;                      // Bytes that were sent on the wire, but not yet acked
     uint64_t PostedBytes;                        // Total bytes queued, but not yet acked. These may contain sent bytes that may have potentially lost too.
     uint64_t IdealBytes;                         // Ideal number of bytes required to be available to  avoid limiting throughput
     uint64_t SmoothedRTT;                        // Smoothed RTT value
-    uint64_t CongestionWindow;                   // Congestion Window
+    uint32_t CongestionWindow;                   // Congestion Window
     uint64_t Bandwidth;                          // Estimated bandwidth
 
 } NETWORK_STATISTICS;
